more haproxy ports for bitcoin
This includes ports for mainnet, testnet and regtest. Of course, not all
three may be running on the current network.

This may prove to be a bad idea. The bitcoin mode could be determined by
an environment variable when the container is launched, and using
different ports just becomes a burden.

fixes #28

original-commit-id: 4169460cc51aedfa667251d4fbd4871a169c01bb